table-facade-jdbc-parent/table-facade-jdbc-mysql/src/test/java/io/github/openfacade/table/jdbc/mysql/MysqlJdbcTableManagementTest.java
/*
 * Copyright 2024 OpenFacade Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package io.github.openfacade.table.jdbc.mysql;

import com.zaxxer.hikari.HikariConfig;
import com.zaxxer.hikari.HikariDataSource;
import io.github.openfacade.table.api.TableException;
import lombok.extern.slf4j.Slf4j;
import org.junit.jupiter.api.*;
import org.testcontainers.containers.MySQLContainer;
import org.testcontainers.junit.jupiter.Container;
import org.testcontainers.junit.jupiter.Testcontainers;

import java.sql.Connection;
import java.sql.Statement;
import javax.sql.DataSource;

import static org.junit.jupiter.api.Assertions.*;

@Slf4j
@Testcontainers
class MysqlJdbcTableManagementTest {

    @Container
    private static final MySQLContainer<?> mysqlContainer = new MySQLContainer<>("mysql:8.0")
            .withDatabaseName("testdb")
            .withUsername("test")
            .withPassword("test");

    private static DataSource dataSource;
    private MysqlJdbcTableManagement tableManagement;

    @BeforeAll
    static void setUp() {
        HikariConfig config = new HikariConfig();
        config.setJdbcUrl(mysqlContainer.getJdbcUrl());
        config.setUsername(mysqlContainer.getUsername());
        config.setPassword(mysqlContainer.getPassword());
        config.setDriverClassName(mysqlContainer.getDriverClassName());

        dataSource = new HikariDataSource(config);
    }

    @BeforeEach
    void init() {
        tableManagement = new MysqlJdbcTableManagement(dataSource);
    }

    @Test
    public void testDropTableSuccess() throws Exception {
        try (Connection connection = dataSource.getConnection();
             Statement stmt = connection.createStatement()) {
            stmt.executeUpdate("CREATE TABLE test_table (id INT PRIMARY KEY)");
        }

        tableManagement.dropTable("test_table");

        try (Connection connection = dataSource.getConnection();
             Statement stmt = connection.createStatement()) {
            stmt.executeQuery("SELECT * FROM test_table");
            fail("Table 'test_table' should have been dropped");
        } catch (Exception e) {
            assertTrue(e.getMessage().contains("doesn't exist"));
        }
    }

    @Test
    public void testDropNotExistTableFail() {
        Exception exception = assertThrows(TableException.class, () -> {
            tableManagement.dropTable("non_existent_table");
        });

        String expectedMessage = "drop table failed";
        String actualMessage = exception.getMessage();

        assertTrue(actualMessage.contains(expectedMessage));
    }
}
